Add a local_manifest style manifest. upstream can look after the tracking while we kick back and drink smashed raspberry lemonade cocktails!!
2 ounces Triple Sec
2 ounces Limoncello
1 handful of raspberries
juice of two lemons
club soda
raw sugar

To muddle your fruit, add a handful of fresh raspberries and a sprinkle of
raw sugar to the bottom of a glass. Use a muddler (a wooden spoon will work too!)
to smash your fruit until it has the appearance of jam.

Add the Triple Sec, Limoncello and lemon juice to your glass. Add ice.
Top if off with club soda. Give it a quick stir and enjoy!
Garnished mine with fresh raspberries and lemon slices!

http://www.abeautifulmess.com/2013/04/smashed-raspberry-lemonade-cocktail.html